pyproject.toml
[project]
name = "earthquacke-map"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "mysql-connector-python>=9.3.0",
    "requests>=2.32.3",
]
